Refactor Offensive Spell Accuracy to float

+++ Assets/Spells/OffensiveSpell.cs
@@ -3,6 +3,6 @@ namespace Assets.Spells
     public abstract class OffensiveSpell: SpellBase
     {
         public abstract int AttackPower { get; }
-        public abstract int Accuracy { get; }
+        public abstract float Accuracy { get; }
     }
 }
